requirements.txt
aiogram==3.15.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.1
pydantic>=2.4.1,<2.10
pydantic-settings>=2.0,<3.0
httpx==0.28.1
google-genai==1.7.0
groq==0.15.0
Pillow==11.1.0
langdetect==1.0.9
